add users, authorization and authentication

+++ src/repositories/scheduledTransactionRepository.ts
@@ -6,25 +6,6 @@ import {
 } from '../types/scheduledTransaction';
 
 class ScheduledTransactionRepository {
-  private mapScheduledTransactionDbToDomain(
-    db: any,
-  ): ScheduledTransactionDomain {
-    return {
-      id: db.id,
-      description: db.description,
-      value: db.value,
-      type: db.type,
-      categoryId: db.categoryId,
-      scheduleType: db.scheduleType,
-      interval: db.interval === null ? undefined : db.interval,
-      dayOfWeek: db.dayOfWeek === null ? undefined : db.dayOfWeek,
-      dayOfMonth: db.dayOfMonth === null ? undefined : db.dayOfMonth,
-      monthOfYear: db.monthOfYear === null ? undefined : db.monthOfYear,
-      lastRunDate: db.lastRunDate === null ? undefined : db.lastRunDate,
-      nextRunDate: db.nextRunDate === null ? undefined : db.nextRunDate,
-    };
-  }
-
   public async createScheduledTransaction(
     data: CreateScheduledTransaction,
     nextRunDate: Date,
@@ -37,15 +18,19 @@ class ScheduledTransactionRepository {
         dayOfMonth: data.dayOfMonth === undefined ? null : data.dayOfMonth,
         monthOfYear: data.monthOfYear === undefined ? null : data.monthOfYear,
         nextRunDate,
+        userId: data.userId,
       },
     });
     return scheduledTransaction.id;
   }
 
-  public async getAllScheduledTransactions(): Promise<
-    ScheduledTransactionDomain[]
-  > {
+  public async getAllScheduledTransactions(
+    userId: string,
+  ): Promise<ScheduledTransactionDomain[]> {
     const scheduledTransactions = await prisma.scheduledTransaction.findMany({
+      where: {
+        userId,
+      },
       orderBy: {
         nextRunDate: 'asc',
       },
@@ -78,10 +63,11 @@ class ScheduledTransactionRepository {
   public async updateScheduledTransaction(
     id: string,
     data: UpdateScheduledTransaction,
+    userId: string,
     nextRunDate: Date,
   ): Promise<string> {
     const scheduledTransaction = await prisma.scheduledTransaction.update({
-      where: { id },
+      where: { id, userId },
       data: {
         ...data,
         interval: data.interval === undefined ? null : data.interval,
@@ -94,21 +80,45 @@ class ScheduledTransactionRepository {
     return scheduledTransaction.id;
   }
 
-  public async deleteScheduledTransaction(id: string): Promise<void> {
-    await prisma.scheduledTransaction.delete({ where: { id } });
+  public async deleteScheduledTransaction(
+    id: string,
+    userId: string,
+  ): Promise<void> {
+    await prisma.scheduledTransaction.delete({ where: { id, userId } });
   }
 
   public async getScheduledTransactionById(
     id: string,
+    userId: string,
   ): Promise<ScheduledTransactionDomain | null> {
     const scheduledTransaction = await prisma.scheduledTransaction.findUnique({
-      where: { id },
+      where: { id, userId },
     });
 
     if (!scheduledTransaction) {
       return null;
     }
     return this.mapScheduledTransactionDbToDomain(scheduledTransaction);
+  }
+
+  private mapScheduledTransactionDbToDomain(
+    db: any,
+  ): ScheduledTransactionDomain {
+    return {
+      id: db.id,
+      description: db.description,
+      value: db.value,
+      type: db.type,
+      categoryId: db.categoryId,
+      scheduleType: db.scheduleType,
+      interval: db.interval === null ? undefined : db.interval,
+      dayOfWeek: db.dayOfWeek === null ? undefined : db.dayOfWeek,
+      dayOfMonth: db.dayOfMonth === null ? undefined : db.dayOfMonth,
+      monthOfYear: db.monthOfYear === null ? undefined : db.monthOfYear,
+      lastRunDate: db.lastRunDate === null ? undefined : db.lastRunDate,
+      nextRunDate: db.nextRunDate === null ? undefined : db.nextRunDate,
+      userId: db.userId,
+    };
   }
 }
